% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R, R/server_connections.R
\name{editcred}
\alias{editcred}
\title{Open the .Renviron File in Editor}
\usage{
editcred(homepath = "~/")

editcred(homepath = "~/")
}
\arguments{
\item{homepath}{the path to the home directory}
}
\value{
open the .Renviron file to add PostgreSQL credentials

open the .Renviron file to add PostgreSQL credentials
}
\description{
This makes it possible to set the credential variables in .Renviron file.
It can be in the user home directory or possibly in the R project working directory.

This makes it possible to set the credential variables in .Renviron file.
It can be in the user home directory or possibly in the R project working directory.
}
