% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_imputedList}
\alias{get_imputedList}
\title{Generate List of Imputed Values per Species}
\usage{
get_imputedList(
  trimIndex = NULL,
  years = 2010:(lubridate::year(lubridate::today()) - 1),
  Art = 1:200,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  indicator_layout = FALSE,
  indicators = NULL,
  indicatorname = NULL,
  verification = c(109, 110, 111),
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  write = FALSE
)
}
\arguments{
\item{trimIndex}{a trim index object from [get_trimindex()]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data from}

\item{Region}{character or reg ex; which region do you want. Possible values are:
'SGot', 'OGot', 'VGotSve', 'OSve', 'NSveNor', 'NNor'.}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{indicator_layout}{logical; whether to use the indicator species and add
nr of sites statistics. If TRUE this overrides `Art`}

\item{indicators}{optional; you can add a concatenated list of indicator
species uids for a new indicator}

\item{indicatorname}{the name of the new indicator. If indicators is given
but without setting name the indicator will be named 'NewInd'}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the data sources}

\item{write}{logical; if index should be written to csv}
}
\value{
a data frame with trim indices per species
}
\description{
Generate List of Imputed Values per Species
}
