% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\name{get_indicatorAnalyses}
\alias{get_indicatorAnalyses}
\title{Run Indicator Analysis}
\usage{
get_indicatorAnalyses(infile = NULL, origin = "Sverige", indicators = "ALL")
}
\arguments{
\item{infile}{list of imputed index from [get_imputedList()]}

\item{origin}{the origin of list species}

\item{indicators}{which indicators to get index from; default to 'ALL'}
}
\description{
Run Indicator Analysis
}
