% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby.R
\name{get_nearby}
\alias{get_nearby}
\title{Get Nearby Places to Given Coordinates}
\usage{
get_nearby(df, radius = 50, top = 1, limited = TRUE, population_limit = 0)
}
\arguments{
\item{df}{data frame with coordinates, or sf object from e.g. [sebms_sunhours_data()]}

\item{radius}{the radius to search for nearby places}

\item{top}{how many of the top results to save}

\item{limited}{logical; if you want only the names of the resulting sites}

\item{population_limit}{the smallest amount of poeple that should be in the
locations}
}
\value{
a data frame with location names nearby your coordinates
}
\description{
Search for a city, village or municipality nearby a coordinate depending on
what is available and the number of inhabitants.
}
