% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby.R
\name{get_nearby_SunHour}
\alias{get_nearby_SunHour}
\title{Get Nearby Places to Sunhour Stations}
\usage{
get_nearby_SunHour(
  df,
  radius = 50,
  top = 1,
  limited = TRUE,
  population_limit = 0,
  sunvar = total_sunH,
  findMaxMin = TRUE
)
}
\arguments{
\item{df}{data frame with coordinates, or sf object}

\item{radius}{the radius to search for nearby places}

\item{top}{how many of the top results to save}

\item{limited}{logical; if you want only the names of the resulting sites}

\item{population_limit}{the smallest amount of poeple that should be in the
locations}

\item{sunvar}{variable with sun hour data`}

\item{findMaxMin}{logical (default TRUE); do you want only the location with
most and least sunhours.}
}
\value{
a data frame with location names nearby your coordinates
}
\description{
Search for a city, village or municipality nearby the Sun Hour stations and
get the Sun hour data. Optional is to get the only stations where there where
most and least number of sun hours. This function also depends on what is
available and the number of inhabitants.
}
