% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trendHistogram}
\alias{get_trendHistogram}
\title{Change of Species Abundance}
\usage{
get_trendHistogram(
  trendIndex = NULL,
  trimIndex = NULL,
  years = 2010:lubridate::year(lubridate::today()) - 1,
  Art = 1:200,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  verification = c(109, 110, 111),
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  write = FALSE,
  print = TRUE,
  indicators = FALSE
)
}
\arguments{
\item{trendIndex}{optional, output from [get_trendIndex()]}

\item{trimIndex}{optional, a trim index object from [get_trimindex()], ignored if
trendIndex is given}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data from}

\item{Region}{character or reg ex; which region do you want. Possible values are:
'SGot', 'OGot', 'VGotSve', 'OSve', 'NSveNor', 'NNor'.}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the data sources}

\item{write}{logical; if index should be written to csv}

\item{indicators}{logical; if TRUE use the indicators as species selection,
this override `Art`}
}
\value{
Two histograms of abundance change in percent, on a linear scale and on a log
  scale. Saved as a png-files, printed or both.
}
\description{
Creates a figure of percent change in species abundances as a histogram with five
change categories.
}
