% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\name{get_trimComparedPlots}
\alias{get_trimComparedPlots}
\title{Create and Save Local TRIM Plots with National TRIM Reference}
\usage{
get_trimComparedPlots(
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  Art = 1:200,
  trimmedImputedSwedishList = NULL,
  year = 2010:lubridate::year(lubridate::today())
)
}
\arguments{
\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{Art}{the species of interest}

\item{trimmedImputedSwedishList}{data frame of national wide data of species
trim values}

\item{year}{the year of interest}
}
\value{
figures saved as png comparing national and local trim indices
}
\description{
Create and Save Local TRIM Plots with National TRIM Reference
}
