% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimIndex}
\alias{get_trimIndex}
\title{Calculate TRIM Index}
\usage{
get_trimIndex(
  infile = NULL,
  years = 2010:lubridate::year(lubridate::today()) - 1,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  ...
)
}
\arguments{
\item{infile}{file with site, year, total observations and reverse frequency weight (1/#visits), or an object from [get_trimInfile()]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{...}{further arguments to pass to [get_trimInfile()]}
}
\value{
a trim file with yearly changes of each species.
}
\description{
Calculate TRIM Index
}
