% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimInfile}
\alias{get_trimInfile}
\title{Create an Object with Species Number per Year and Site with Visit Frequency}
\usage{
get_trimInfile(
  years = 2010:2023,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  filterPattern = NULL,
  topList = FALSE,
  topNumber = 200,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{filterPattern}{a regex pattern to filter SQL query}

\item{topList}{logical; whether the top list of species should be used}

\item{topNumber}{the number of top most observed species}

\item{source}{the data sources}
}
\value{
a tibble with site, year, as well as the number of individuals
  observed and the observation frequency for that year, species, and site.
}
\description{
Create a tibble with species ID and name together with the total number of observations of each species and the frequency (1 / #visits)
}
