% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimPlots}
\alias{get_trimPlots}
\title{Create and Save TRIM Plots}
\usage{
get_trimPlots(trimIndex = NULL, years = 2010:2023, Art = 1:200, ...)
}
\arguments{
\item{trimIndex}{optional; a trimIndex object from the [get_trimIndex()]}

\item{years}{the years to calculate trim index on, ignored if a trimIndex file is given}

\item{Art}{the species of interest, ignored if trimIndex is not NULL}

\item{...}{optional; other arguments passed on to [get_trimInfile()]}
}
\value{
figures in png format of the species trends with confidence interval
}
\description{
Create and Save TRIM Plots
}
