% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\docType{data}
\name{indicatorlist}
\alias{indicatorlist}
\title{Indicator Species for Trim Index}
\format{
An object of class \code{list} of length 4.
}
\usage{
indicatorlist
}
\value{
list of species uids for 'grassland', 'agricultural', 'forest', and
  'common20' species
}
\description{
Indicator Species for Trim Index
}
\keyword{datasets}
