% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_abundance_year_compare_plot}
\alias{sebms_abundance_year_compare_plot}
\title{Butterfly Number Histogram  Plot}
\usage{
sebms_abundance_year_compare_plot(
  year = 2021:2022,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE
)
}
\description{
Show the number of found butterflies per week, compared between two years.
}
