% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherplots.R
\name{sebms_ggsave}
\alias{sebms_ggsave}
\title{Saves a ggplot object as a PNG file, resizing using pixel dimensions and a text scaling factor}
\usage{
sebms_ggsave(
  plot,
  filename,
  width = 12.67,
  height = 9.722,
  text.factor = 3,
  weathervar = "Temp"
)
}
\arguments{
\item{plot}{a ggplot object}

\item{filename}{the path to the output file}

\item{width}{pixel width}

\item{height}{pixel height}

\item{text.factor}{text scaling factor (default is 3)}

\item{weathervar}{which weather variable it shoulld put in the name; 'Temp' or 'Precip'}
}
\description{
Saves a ggplot object as a PNG file, resizing using pixel dimensions and a text scaling factor
}
