% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_local_transekt_map}
\alias{sebms_local_transekt_map}
\title{Create Local Maps with Transect}
\usage{
sebms_local_transekt_map(
  year = lubridate::year(lubridate::today()) - 1,
  occ_sp,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  width = 12,
  height = 18,
  maptype = "both",
  print = FALSE,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{maptype}{what survey type to produce map on, can be 'Transect', 'Point'
or 'both', #'   default to 'both'. The 'Transect' and 'Point' can be
abbreviated to 'T' and #'   'P'.}

\item{print}{logical; should the plots be printed in window, default FALSE}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a png file with a map of the chosen area with slingor or transects marked.
}
\description{
Creates a map of the County or Municipality with the transect marked.
}
