% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soltimmar.R
\name{sebms_minmax_sunhour}
\alias{sebms_minmax_sunhour}
\title{Maximum and Minimum Sun-hours Given Years}
\usage{
sebms_minmax_sunhour(df, years = 2017:2022, month = 4:9)
}
\arguments{
\item{df}{a sf object with `year` and `total_sunhour` created by `sebms_sunhour_data()`}

\item{years}{optional; the years to create the min max for}

\item{month}{optional; the month to summarise the sunhours over}
}
\value{
a data frame with the max and min of total sunhours per year and the mean and diff from mean at that lokation. It also gives the name of the nearest city or village for that location.
}
\description{
GIve the maximm and minimum sunhour per year and the city or village closest to that location
}
