% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_occurances_distribution}
\alias{sebms_occurances_distribution}
\title{Retrieve Species Abundance List per Site with Coordinates}
\usage{
sebms_occurances_distribution(
  year = 2020:2021,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with species ids and names, filtered for trimmed species
  names, the site ids and names, site type, max number of individuals observed for each siote,
  the date, county, region, and municipality, and the rank of site
}
\description{
Returns a data frame of counts per species and site with their coordinates in
SWEREF. This is used for the [sebms_distribution_map()]
}
