% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_ggtheme.R
\docType{data}
\name{sebms_palette}
\alias{sebms_palette}
\title{Palette Used in ggplots for Weather Plots}
\format{
An object of class \code{character} of length 2.
}
\usage{
sebms_palette
}
\value{
vector of color hex codes
}
\description{
This is also used in some of the trim plots.
}
\keyword{datasets}
