% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_precip_temp_plot}
\alias{sebms_precip_temp_plot}
\title{Plot of temperature and precipitation data}
\usage{
sebms_precip_temp_plot(filter_cities, df_precip, df_temp)
}
\arguments{
\item{filter_cities}{a vector of cities, if missing a default of c("Umeå", "Stockholm", "Visby", "Lund") is used}

\item{df_precip}{a data frame with precipitation data in a specific format, see vignette for details, default uses packaged data sebms_data_precip_2015}

\item{df_temp}{a data frame with temperature data in a the same format, see vignette for details, default uses packaged data sebms_data_temp_2015}
}
\value{
a ggplot plot object
}
\description{
Plot of temperature and precipitation data
}
