% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_precip_temp_png}
\alias{sebms_precip_temp_png}
\title{Creates png Figures  of Temperature and Precipitation for each Site}
\usage{
sebms_precip_temp_png(my_place = NA)
}
\arguments{
\item{my_place}{the places you want weather data pngs from (default to Umeå, Stockholm, Visby, Lund)}
}
\value{
a png file
}
\description{
Creates png Figures  of Temperature and Precipitation for each Site
}
