% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_regional_site_map}
\alias{sebms_regional_site_map}
\title{Create Local Maps with Transect and Point Locals}
\usage{
sebms_regional_site_map(
  year = lubridate::year(lubridate::today()) - 1,
  occ_sp,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  active_site_cutoff = NULL,
  width = 12,
  height = 18,
  zoomlevel = NULL,
  maptype = "both",
  showgrid = F,
  print = FALSE,
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{zoomlevel}{the level of zoom on map. This is set automatically but this argument
allow for changing this if wanted.}

\item{maptype}{what survey type to produce map on, can be 'Transect', 'Point', 'Transekt', 'Punkt',
or 'both' (default). The 'Transect' and 'Point' can be
abbreviated to 'T' and 'P'.}

\item{showgrid}{show the grid lines on the map.}

\item{print}{logical; should the plots be printed in window, default FALSE}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}
}
\value{
one or two png files with a map of the chosen area with slingor or transects
  marked.
}
\description{
Creates a map of the County or Municipality with the transect and point data
marked.You need to be on a Lund university or SLU network, or LU/SLU VPN to get the map
as it is created from the SLU WMS topomap.
}
