% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_distribution.R
\name{sebms_sites_map}
\alias{sebms_sites_map}
\title{Swedish Map of SeBMS Sites}
\usage{
sebms_sites_map(year = 2021, width = 5, height = 3.5)
}
\arguments{
\item{year}{the year of interest}

\item{width}{the plot width, default 5 inches}

\item{height}{the plot height, default 3.5 inches}
}
\value{
ggplot object
}
\description{
Producing map for all sites on Swedish grid.
}
