% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_abundance_plot}
\alias{sebms_species_abundance_plot}
\title{Species Histogram Plot}
\usage{
sebms_species_abundance_plot(
  year = 2021,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  plotname = FALSE,
  database = TRUE
)
}
\arguments{
\item{year}{year to use for plot}

\item{Art}{integer; the species id of interest}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{plotname}{logical; if you want the species name inside the plot}

\item{database}{logical; if the data should be based on the sebms database}
}
\value{
A png per species showing the number of individuals per week.
}
\description{
Show the number of individuals per week of a given species and year.
}
