% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_count}
\alias{sebms_species_count}
\title{Retrieve Total Species List from SeBMS}
\usage{
sebms_species_count(year = 2021:2022)
}
\description{
Species counts per date for given year over all sites.
This is used eg for the histograms
}
