% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_count_filtered}
\alias{sebms_species_count_filtered}
\title{Retrieve Species Abundance List per Date}
\usage{
sebms_species_count_filtered(
  year = 2020:2021,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\description{
Returns a data frame with counts per species and date.
This is used in most plot functions:
`sebms_abundance_per_species_plot()`, `sebms_abundance_year_compare_plot()`,
and `sebms_species_abundance_plot()`
}
