% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_histo_plot}
\alias{sebms_species_histo_plot}
\title{Species Histogram Plot}
\usage{
sebms_species_histo_plot(year = 2021, Art = "Luktgräsfjäril", database = TRUE)
}
\arguments{
\item{year}{year to use for plot}

\item{Art}{The species of interest}

\item{database}{logical; if the data should be based on the sebms database}
}
\description{
Show the number of individuals per week of a given species and year.
}
