% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_histo_plot_orig}
\alias{sebms_species_histo_plot_orig}
\title{Species histo plot - original version}
\usage{
sebms_species_histo_plot_orig()
}
\description{
This plot should probably be parameterized with the week range (defaulted to 13..42) used for the x-axis and with the relevant year context, also the range of y-axis should not be hardcoded in the future.
}
