% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_per_sitetype_plot}
\alias{sebms_species_per_sitetype_plot}
\title{Species per Site and Site Type Plot}
\usage{
sebms_species_per_sitetype_plot(database = TRUE)
}
\description{
Species per Site and Site Type Plot
}
