% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_per_year}
\alias{sebms_species_per_year}
\title{Retrieve Species List per Year}
\usage{
sebms_species_per_year()
}
\description{
Give a list of all species counts from all databases for each year
}
