% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_species_per_year_filtered}
\alias{sebms_species_per_year_filtered}
\title{Retrieve Species Abundance List per Year}
\usage{
sebms_species_per_year_filtered(year = 2020:2021, verification = 109)
}
\arguments{
\item{year}{year or years to use for plot}

\item{verification}{a verification code that filter out verified occurenses of species, default to 109.}
}
\description{
Return a data frame with abundance data per species and year.
}
