% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soltimmar.R
\name{sebms_sunhours_data}
\alias{sebms_sunhours_data}
\title{Create a total Sunhours Dataframe from SMHI Sunhour data}
\usage{
sebms_sunhours_data(year = year(today()) - 1, month = 4:9)
}
\arguments{
\item{year}{the year to produce plot for}

\item{month}{numeric value of the months to summarise sun ours over (default to 4:9)}
}
\value{
a sf spatial point object with the WGS84 coordinate system
}
\description{
Produce a data frame of the total irradiance in Sweden for the given month.
}
