% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\docType{data}
\name{sebms_trimpal}
\alias{sebms_trimpal}
\title{Palette Used in ggplots for Trim Index}
\format{
An object of class \code{character} of length 3.
}
\usage{
sebms_trimpal
}
\value{
vector of color hex codes
}
\description{
Palette Used in ggplots for Trim Index
}
\keyword{datasets}
