% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_trimSites}
\alias{sebms_trimSites}
\title{Retrieve Sites for Trim Functions}
\usage{
sebms_trimSites(
  year = 2010:lubridate::year(lubridate::today()),
  Landskap = ".",
  Region = ".",
  Län = ".",
  Kommun = ".",
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Region}{character or reg ex; which bioregion you want the data from}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with species IDs, name, and min and max flight time week
}
\description{
This function retrieve the sites used for the trim indices
}
