% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_trimSpecies}
\alias{sebms_trimSpecies}
\title{Retrieve Species Data for Trim Functions}
\usage{
sebms_trimSpecies(
  year = 2010:lubridate::year(lubridate::today()),
  Art = 1:200,
  filterPattern = NULL,
  topList = FALSE,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{filterPattern}{a regex pattern to filter SQL query}

\item{topList}{logical; whether the top list of species should be used}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with species IDs, name, and min and max flight time week
}
\description{
This function retrieve min and max flight week for each species. used for the
trim indices
}
