% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_trimvisits}
\alias{sebms_trimvisits}
\title{Get Visits for Trim Indices}
\usage{
sebms_trimvisits(
  year = 2010:lubridate::year(lubridate::today()),
  minmax = 22:32,
  source = c(54, 55, 56, 63, 64, 66, 67)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{minmax}{the first and last week of interest}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with visits per year and site.
}
\description{
Retrieve data with visits per site per year.
}
