% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_ggtheme.R
\name{theme_sebms}
\alias{theme_sebms}
\title{Theme for Swedish Butterfly Monitoring Scheme for Weather Figures}
\usage{
theme_sebms(
  title_sz = 24,
  x_title_sz = 14,
  y_title_sz = 14,
  x_sz = 12,
  y_sz = 12,
  legend_position = "none",
  fontfamily = "Arial"
)
}
\arguments{
\item{title_sz}{font size for title}

\item{x_title_sz}{font size for x axis title}

\item{y_title_sz}{font size for y axis title}

\item{x_sz}{font size for x axis text}

\item{y_sz}{font size for y axis text}

\item{legend_position}{char indicating legend_position such as "none"}
}
\value{
a theme that can be used for ggplot plot objects
}
\description{
Theme for Swedish Butterfly Monitoring Scheme for Weather Figures
}
