% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\name{get_imputedList}
\alias{get_imputedList}
\title{Generate List of Imputed Values per Species}
\usage{
get_imputedList(
  trimIndex = NULL,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  indicator_layout = FALSE,
  year = 2010:lubridate::year(lubridate::today()),
  ...
)
}
\arguments{
\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{indicator_layout}{logical; whether the list should contain the}

\item{year}{the year of interest}

\item{...}{extra filter parameters passed to the [trimInfile()] function}

\item{origin}{the origin of list species}
}
\value{
a data frame with trim indices per species
}
\description{
Generate List of Imputed Values per Species
}
