% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_imputedList}
\alias{get_imputedList}
\title{Generate List of Imputed Values per Species}
\usage{
get_imputedList(
  trimIndex = NULL,
  years = 2010:lubridate::year(lubridate::today()),
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  indicator_layout = FALSE,
  ...
)
}
\arguments{
\item{trimIndex}{a trim index object from [get_trimindex()]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{indicator_layout}{logical; whether the list should contain the}

\item{...}{extra filter parameters passed to the [trimInfile()] function}
}
\value{
a data frame with trim indices per species
}
\description{
Generate List of Imputed Values per Species
}
