% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_indicatorAnalyses}
\alias{get_indicatorAnalyses}
\title{Run Indicator Analysis}
\usage{
get_indicatorAnalyses(
  infile = NULL,
  years = 2010:lubridate::year(lubridate::today()) - 1,
  lastyear = 7,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  write = TRUE,
  print = FALSE,
  indicators = NULL,
  indicatorname = NULL
)
}
\arguments{
\item{infile}{list of imputed index from [get_imputedList(indicator_layout =
TRUE)]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{lastyear}{number of years to do a trend comparison with}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{write}{logical; if you want to write result to csv files, default TRUE.}

\item{print}{logical; if you want to print result to output, default FALSE}

\item{indicators}{optional; you can add a concatenated list of indicator
species uids for a new indicator}

\item{indicatorname}{the name of the new indicator. If indicators is given
but without setting name the indicator will be named 'NewInd'}
}
\value{
two csv files for each indicator groups. One with indicator index and
  changes and one with trend data.
}
\description{
Run Indicator Analysis
}
