% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_indicatorPlots}
\alias{get_indicatorPlots}
\title{Create and Save Indicator Trend Plots}
\usage{
get_indicatorPlots(
   msi_out = NULL,
   years = 2010:2023,
   write = TRUE,
   print = FALSE
)
}
\arguments{
\item{msi_out}{output from [get_indicatorAnalyses()]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{write}{logical; if you want to write result to csv files, default TRUE.}

\item{print}{logical; if you want to print result to output, default FALSE}
}
\value{
trend plots with confidence interval for the indicator groups, saved
  as png files.
}
\description{
Create and Save Indicator Trend Plots
}
