% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_indicatorPlots}
\alias{get_indicatorPlots}
\title{Create and Save Indicator Trend Plots}
\usage{
get_indicatorPlots(
   msi_out = NULL,
   years = 2010:2023,
   Län = ".",
   Landskap = ".",
   Kommun = ".",
   write = TRUE,
   print = FALSE
)
}
\arguments{
\item{msi_out}{output from [get_indicatorAnalyses()]}

\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the data sources}

\item{write}{logical; if you want to write result to csv files, default TRUE.}

\item{print}{logical; if you want to print result to output, default FALSE}
}
\value{
trend plots with confidence interval for the indicator groups, saved
  as png files.
}
\description{
Create and Save Indicator Trend Plots
}
