% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimComparedPlots}
\alias{get_trimComparedPlots}
\title{Create and Save Local TRIM Plots with National TRIM Reference}
\usage{
get_trimComparedPlots(
  years = 2010:lubridate::year(lubridate::today()) - 1,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  trimmedImputedSwedishList = NULL,
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  write = TRUE,
  print = TRUE
)
}
\arguments{
\item{years}{the years to calculate trim index on, ignored if a trimIndex file is given}

\item{Art}{the species of interest, ignored if trimIndex is not NULL}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{trimmedImputedSwedishList}{data frame of national wide data of species
trim values}

\item{source}{the data sources}

\item{write}{logical; if you want to write result to csv files, default TRUE}

\item{print}{logical; if you want to print result to output, default FALSE}
}
\value{
figures saved as png comparing national and local trim indices
}
\description{
Create and Save Local TRIM Plots with National TRIM Reference
}
