% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimComparedPlots}
\alias{get_trimComparedPlots}
\title{Create and Save Local TRIM Plots with National TRIM Reference}
\usage{
get_trimComparedPlots(
  years = 2010:lubridate::year(lubridate::today()),
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  trimmedImputedSwedishList = NULL
)
}
\arguments{
\item{years}{the year span of interest, set as 'firstyear:lastyear'.}

\item{Art}{the species of interest}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{trimmedImputedSwedishList}{data frame of national wide data of species
trim values}
}
\value{
figures saved as png comparing national and local trim indices
}
\description{
Create and Save Local TRIM Plots with National TRIM Reference
}
