% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\name{get_trimIndex}
\alias{get_trimIndex}
\title{Calculate TRIM Index}
\usage{
get_trimIndex(infile = NULL, year = 2010:2023, Art = 1:200, ...)
}
\arguments{
\item{infile}{file with site, year, total observations and reverse frequency weight (1/#visits), or an object from [get_trimInfile()]}

\item{year}{the year of interest}

\item{Art}{the species of interest}

\item{...}{further arguments to pass to [get_trimInfile()]}
}
\value{
a trim file with yearly changes of each species.
}
\description{
Calculate TRIM Index
}
