% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_trimfunctions.R
\name{get_trimPlots}
\alias{get_trimPlots}
\title{Create and Save TRIM Plots}
\usage{
get_trimPlots(
  trimIndex = NULL,
  year = 2010:2023,
  Art = 1:200,
  path = getwd(),
  ...
)
}
\arguments{
\item{trimIndex}{optional; a trimIndex object from the [get_trimIndex()]}

\item{year}{the years to calculate trimindex on, ignored if trimIndex is not
NULL}

\item{Art}{the species of interest, ignored if trimIndex is not NULL}

\item{path}{the path to where to save figures}

\item{...}{optional; other arguments passed on to [get_trimInfile()]}
}
\value{
figures in png format of the species trends with confidence interval
}
\description{
Create and Save TRIM Plots
}
