% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimfunctions.R
\name{get_trimPlots}
\alias{get_trimPlots}
\title{Create and Save TRIM Plots}
\usage{
get_trimPlots(
  trimIndex = NULL,
  years = 2010:lubridate::year(lubridate::today()) - 1,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  xaxis_sep = 5,
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  write = TRUE,
  print = TRUE,
  ...
)
}
\arguments{
\item{trimIndex}{optional; a trimIndex object from the [get_trimIndex()]}

\item{years}{the years to calculate trim index on, ignored if a trimIndex file is given}

\item{Art}{the species of interest, ignored if trimIndex is not NULL}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the data sources}

\item{write}{logical; if you want to write result to csv files, default TRUE}

\item{print}{logical; if you want to print result to output, default FALSE}

\item{...}{optional; other arguments passed on to [get_trimInfile()]}
}
\value{
figures in png format of the species trends with confidence interval
}
\description{
Create and Save TRIM Plots
}
