% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_abundance_per_species_plot}
\alias{sebms_abundance_per_species_plot}
\title{Individuals per Species Plots}
\usage{
sebms_abundance_per_species_plot(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year to use for plot}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{database}{logical; if the data should be based on the sebms database}

\item{source}{the database sources as id numbers}
}
\value{
Two png figures with the abundance data for each species.
One for species count below the median for that year and one for species above median.
}
\description{
Produce a plot with number of individuals per species a given year
}
