% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_abundance_per_species_plot}
\alias{sebms_abundance_per_species_plot}
\title{Cumulative Species List Plots}
\usage{
sebms_abundance_per_species_plot(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE
)
}
\arguments{
\item{year}{year to use for plot}

\item{Län}{character; which county you want the data from}

\item{Landskap}{character; which region you want the data from}

\item{Kommun}{character; which municipality you want the data from}

\item{database}{logical; if the data should be based on the sebms database}
}
\value{
two png figures with the abundance data for each species.
One for species count below the median for that year and one for above median.
}
\description{
Produce a plot with number of individuals per species a given year
}
