% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesplots.R
\name{sebms_abundance_per_species_plot}
\alias{sebms_abundance_per_species_plot}
\title{Individuals per Species Plots}
\usage{
sebms_abundance_per_species_plot(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  print = FALSE
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{filepath}{a path to a folder where the plot file should be saved}

\item{tag}{a tag to be added at the end of the file name; optional. For instance to
show the county the plot or data contain.}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}

\item{print}{logical; if FALSE (default) the function does not print to plot window.}
}
\value{
Two png figures with the abundance data for each species. One for species count
  below the median for that year and one for species above median.
}
\description{
Produce a plot with number of individuals per species a given year
}
