% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesplots.R
\name{sebms_abundance_year_compare_plot}
\alias{sebms_abundance_year_compare_plot}
\title{Butterfly Number Two Year Comparison Plot}
\usage{
sebms_abundance_year_compare_plot(
  year = 2021:2022,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  verification = c(109, 110, 111),
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  print = FALSE
)
}
\arguments{
\item{year}{two years to compare, e.g. 2021:2022}

\item{Län}{character or regular expression; which county you want the data from}

\item{Region}{character or reg ex; which region do you want. Possible values are:
'SGot', 'OGot', 'VGotSve', 'OSve', 'NSveNor', 'NNor'.}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{filepath}{a path to a folder where the plot file should be saved}

\item{tag}{a tag to be added at the end of the file name; optional. For instance to
show the county the plot or data contain.}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}

\item{print}{logical; if FALSE (default) the function does not print to plot window.}
}
\value{
A png figure with the number of individuals found each of the
  comparing years per week,
}
\description{
Show the number of found butterflies per week, compared between two years.
}
