% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_abundance_year_compare_plot}
\alias{sebms_abundance_year_compare_plot}
\title{Butterfly Number Two Year Comparison Plot}
\usage{
sebms_abundance_year_compare_plot(
  year = 2021:2022,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE
)
}
\arguments{
\item{year}{two years to compare, e.g. 2021:2022}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{database}{logical; if the data should be based on the sebms database}
}
\value{
A png figure with the number of individuals found each of the comparing years per week,
}
\description{
Show the number of found butterflies per week, compared between two years.
}
