% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_connections.R
\name{sebms_assert_connection}
\alias{sebms_assert_connection}
\title{Connection pool used for db connections}
\usage{
sebms_assert_connection(pool, quiet = FALSE)
}
\arguments{
\item{pool}{the connection pool}
}
\description{
Connection pool used for db connections
}
