% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_distribution.R
\name{sebms_distribution_map}
\alias{sebms_distribution_map}
\title{Swedish Map of SeBMS Distribution Data}
\usage{
sebms_distribution_map(
  year = 2021,
  species = 118,
  width = 5,
  height = 3.5,
  occ_sp
)
}
\arguments{
\item{year}{the year of interest}

\item{species}{the species of interest as a species id}

\item{width}{the plot width, default 5 inches}

\item{height}{the plot height, default 3.5 inches}

\item{occ_sp}{SpatialPoints with occurrence data}
}
\value{
ggplot object
}
\description{
Producing distribution map for species on the Swedish grid.
}
