% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebms_distribution.R
\name{sebms_distribution_map}
\alias{sebms_distribution_map}
\title{Swedish Map of SeBMS Distribution Data}
\usage{
sebms_distribution_map(
  year = lubridate::year(lubridate::today()),
  Art = 1:200,
  width = 9,
  height = 18,
  occ_sp,
  print = FALSE
)
}
\arguments{
\item{year}{the year of interest}

\item{Art}{the species of interest as a species id}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{print}{logical; should the plots be printed in window, default FALSE}
}
\value{
ggplot object of map with grid coloured by local density and with
  species occurence points.
}
\description{
Producing distribution map for species on the Swedish grid.
}
