% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_distribution_map}
\alias{sebms_distribution_map}
\title{Swedish Map of SeBMS Distribution Data}
\usage{
sebms_distribution_map(
  year = lubridate::year(lubridate::today()) - 1,
  occ_sp,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  width = 9,
  height = 18,
  print = FALSE,
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{Art}{the species of interest as a species id}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{print}{logical; should the plots be printed in window, default FALSE}

\item{verification}{a verification code that filter out verified occurenses of species, default to 109.}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
ggplot object of map with grid coloured by local density and with
  species occurrence points.
}
\description{
Producing distribution map for species on the Swedish grid.
}
