% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherplots.R
\docType{data}
\name{sebms_palette}
\alias{sebms_palette}
\title{Palette Used in ggplots}
\format{
An object of class \code{character} of length 2.
}
\usage{
sebms_palette
}
\value{
vector of color hex codes
}
\description{
Palette Used in ggplots
}
\keyword{datasets}
