% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_regional_site_map}
\alias{sebms_regional_site_map}
\title{Create Local Maps with Transect and Point Locals}
\usage{
sebms_regional_site_map(
  year = lubridate::year(lubridate::today()) - 1,
  occ_sp,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  filepath = getwd(),
  tag = NULL,
  active_site_cutoff = NULL,
  width = 12,
  height = 18,
  zoomlevel = NULL,
  maptype = "both",
  showsite = TRUE,
  showgrid = FALSE,
  gridtype = "10",
  onemap = FALSE,
  print = FALSE,
  write = TRUE,
  verification = c(109, 110, 111),
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{Län}{character or regular expression; which county you want the data from}

\item{Region}{character or reg ex; which region do you want. Possible values are:
'SGot', 'OGot', 'VGotSve', 'OSve', 'NSveNor', 'NNor'.}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{filepath}{a path to a folder where the plot file should be saved}

\item{tag}{a tag to be added at the end of the file name; optional. For instance to
show the county the plot or data contain.}

\item{active_site_cutoff}{the year where the site is considered old, and get a smaller
point size on the map (default to no year, all are active).}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{zoomlevel}{the level of zoom on map. This is set automatically but this argument
allow for changing this if wanted.}

\item{maptype}{what survey type to produce map on, can be 'Transect', 'Point', 'Transekt', 'Punkt',
or 'both' (default). The 'Transect' and 'Point' can be
abbreviated to 'T' and 'P'.}

\item{showsite}{logical; whether to show the sites on the map}

\item{showgrid}{logical; should the grid show, defaults to FALSE}

\item{gridtype}{which grid to show on the map, can take the value of 'square5',
'square10', or 'hex', can be shortened to '5','10', and 'h'. Ignored if showgrid =
FALSE}

\item{onemap}{logical; wheter to make one map with all sites or ekse two maps with
points and transects separately.}

\item{print}{logical; should the plots be printed in window, default FALSE}

\item{write}{logical; should the file be written to a png, default to TRUE}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}
}
\value{
one or two png files with a map of the chosen area with points or transects
  marked.
}
\description{
Creates a map of the County, Region, Province or Municipality with the transect and
point data marked. You need to be on a Lund university or SLU network, or LU/SLU VPN to
get the map as it is created from the SLU WMS topomap.
}
