% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_sites_map}
\alias{sebms_sites_map}
\title{Swedish Map of SeBMS Sites}
\usage{
sebms_sites_map(
  year = lubridate::year(lubridate::today()) - 1,
  occ_sp,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  width = 12,
  height = 18,
  maptype = "both",
  filepath = getwd(),
  tag = NULL,
  print = FALSE,
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{maptype}{what survey type to produce map on, can be 'Transect', 'Point', 'Transekt', 'Punkt',
or 'both' (default). The 'Transect' and 'Point' can be
abbreviated to 'T' and 'P'.}

\item{filepath}{a path to a folder where the plot file should be saved}

\item{tag}{a tag to be added at the end of the file name; optional. For instance to
show the county the plot or data contain.}

\item{print}{logical; should the plots be printed in window, default FALSE}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}
}
\value{
Figures in png for points, and transects the given year
}
\description{
Producing map for all visited sites on Swedish grid separating transects and
points.
}
