% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_maps.R
\name{sebms_sites_map}
\alias{sebms_sites_map}
\title{Swedish Map of SeBMS Sites}
\usage{
sebms_sites_map(
  year = lubridate::year(lubridate::today()),
  width = 12,
  height = 18,
  occ_sp,
  print = FALSE
)
}
\arguments{
\item{year}{the year of interest}

\item{width}{the plot width, default 12 inches}

\item{height}{the plot height, default 18 inches}

\item{occ_sp}{SpatialPoints with occurrence data}

\item{print}{logical; should the plots be printed in window, default FALSE}
}
\value{
Figures in png for points, and transects the given year
}
\description{
Producing map for all visited sites on Swedish grid separating transects and
points.
}
