% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_abundance_plot}
\alias{sebms_species_abundance_plot}
\title{Species Histogram Plot}
\usage{
sebms_species_abundance_plot(
  year = 2021,
  Art = 1:200,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE
)
}
\arguments{
\item{Art}{The species id of interest}
}
\description{
Show the number of individuals per week of a given species and year.
}
