% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_count_filtered}
\alias{sebms_species_count_filtered}
\title{Retrieve Filtered Species List per Year}
\usage{
sebms_species_count_filtered(year = 2020:2021)
}
\description{
Retrieve Filtered Species List per Year
}
