% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_histo_plot}
\alias{sebms_species_histo_plot}
\title{Species Histogram Plot}
\usage{
sebms_species_histo_plot(year = 2021, Art = "Luktgräsfjäril", database = TRUE)
}
\description{
Show the number of individuals per week of a given species and year.
}
