% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_per_sitetype_plot}
\alias{sebms_species_per_sitetype_plot}
\title{Number of Sites per Species Number Category and Site Type}
\usage{
sebms_species_per_sitetype_plot(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  database = TRUE,
  source = c(54, 55, 56, 63, 64, 66, 67)
)
}
\arguments{
\item{year}{year to use for plot}

\item{Län}{character or regular expression; which county you want the data from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{database}{logical; if the data should be based on the sebms database}

\item{source}{the database sources as id numbers}
}
\value{
A png with the number of sites within each category of number of species, for both slingor and transects. It also shows the mean number of species per site type.
}
\description{
Show the number of sites within a range of species richness found at the site. Also show the mean number of species per site in each site type.
}
