% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesplots.R
\name{sebms_species_per_sitetype_plot}
\alias{sebms_species_per_sitetype_plot}
\title{Number of Sites per Species Number Category and Site Type}
\usage{
sebms_species_per_sitetype_plot(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  verification = 109,
  source = c(54, 55, 56, 63, 64, 66, 67, 84),
  print = FALSE
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Län}{character or regular expression; which county you want the data
from}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{verification}{a verification code that filter out verified occurenses of species, default to 109.}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}

\item{print}{logical; if FALSE (default) the function does not print to plot window.}
}
\value{
A png with the number of sites within each category of number of
  species, for both slingor and transects. It also shows the mean number of
  species per site type.
}
\description{
Show the number of sites within a range of species richness found at the
site. Also show the mean number of species per site in each site type.
}
