% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_species_per_sitetype_plot}
\alias{sebms_species_per_sitetype_plot}
\title{Species per Site and Site Type Plot}
\usage{
sebms_species_per_sitetype_plot(
  year = 2021,
  Län = "",
  Landskap = "",
  Kommun = "",
  database = TRUE
)
}
\description{
Show the number of sites within a range of species number found at the site. Also show the mean number of species per site in each site type.
 
 @inheritParams sebms_specieslist_cum_plots
}
