% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_site_count}
\alias{sebms_species_site_count}
\title{Retrieve Total Species List per Site from SeBMS}
\usage{
sebms_species_site_count(year = 2021)
}
\description{
This is used eg for the histograms
}
