% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_site_count_filtered}
\alias{sebms_species_site_count_filtered}
\title{Retrieve Total Species List per Site from SeBMS}
\usage{
sebms_species_site_count_filtered(
  year = 2021,
  Län = "",
  Landskap = "",
  Kommun = ""
)
}
\arguments{
\item{year}{the year of interest}
}
\description{
Gives a list of species counts per site and date.
This is used eg for the histograms
}
