% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sebms_species_site_count_filtered}
\alias{sebms_species_site_count_filtered}
\title{Retrieve Species Abundance List per Site, Site type, and Date}
\usage{
sebms_species_site_count_filtered(
  year = 2021,
  Län = ".",
  Landskap = ".",
  Kommun = ".",
  source = c(54, 55, 56, 63, 64, 66, 67)
)
}
\arguments{
\item{year}{the year of interest}
}
\description{
Returns a data frame of counts per species, site, site type, and date.
This is used for the `sebms_species_per_sitetype_plot()`
}
