% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_species_site_count_filtered}
\alias{sebms_species_site_count_filtered}
\title{Retrieve Species Abundance List per Site, Site type, and Date}
\usage{
sebms_species_site_count_filtered(
  year = 2021,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  verification = c(109, 110, 111),
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Län}{character or regular expression; which county you want the data from}

\item{Region}{character or reg ex; which region do you want. Possible values are:
'SGot', 'OGot', 'VGotSve', 'OSve', 'NSveNor', 'NNor'.}

\item{Landskap}{character or reg ex; which region you want the data from}

\item{Kommun}{character or reg ex; which municipality you want the data from}

\item{verification}{a verification code that filter out verified occurenses of species,
default to 109.}

\item{source}{the database sources as id numbers, defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with species ids and names, filtered for trimed species
  names, the site ids and names, site type, number of individuals for eahc
  species, the date, county, region, and municipality
}
\description{
Returns a data frame of counts per species, site, site type, and date. This
is used for the [sebms_species_per_sitetype_plot()]
}
