% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_specieslist_cum_plots}
\alias{sebms_specieslist_cum_plots}
\title{Cumulative Species List Plots}
\usage{
sebms_specieslist_cum_plots(
  year = 2021,
  Län = "",
  Landskap = "",
  Kommun = "",
  database = TRUE
)
}
\arguments{
\item{year}{year to use for plot}

\item{Län}{character; which county you want the data from}

\item{Landskap}{character; which region you want the data from}

\item{Kommun}{character; which municipality you want the data from}

\item{database}{logical; if the data should be based on the sebms database}
}
\value{
a list with two ggplot objects, named p1 and p2
}
\description{
Produce a plot with number of individuals per species a given year
}
