% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_specieslist_cum_plots}
\alias{sebms_specieslist_cum_plots}
\title{Cumulative Species List Plots}
\usage{
sebms_specieslist_cum_plots(database = TRUE)
}
\value{
a list with two ggplot objects, named p1 and p2
}
\description{
Cumulative Species List Plots
}
