% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soltimmar.R
\name{sebms_sundiff_plot}
\alias{sebms_sundiff_plot}
\title{Create a Figure that Shows the Difference in Sun Hours}
\usage{
sebms_sundiff_plot(
  year = lubridate::year(lubridate::today()) - 1,
  df,
  months = 4:9,
  per_month = FALSE,
  legends = FALSE,
  per_day = FALSE
)
}
\arguments{
\item{year}{the year or years to produce plot for}

\item{df}{optinal; dataframe from the [sebms_sunhour_diff()]}

\item{months}{numeric value of the months to summarise sun ours over (default
to 4:9)}

\item{per_month}{logical; summarise per month instead of per year}

\item{legends}{logical; if you want a legend to the figures (default: FALSE)}

\item{per_day}{logical; if data should be downloaded per day}
}
\value{
a figure that shows diffeence in sunhours
}
\description{
Produce a plot that shows differences in sun hours between a given year and a
5-year mean.
}
