% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soltimmar.R
\name{sebms_sundiff_plot}
\alias{sebms_sundiff_plot}
\title{Create a Figure that Shows the Difference in Sunhours}
\usage{
sebms_sundiff_plot(
  year = year(today()) - 1,
  df,
  month = 4:9,
  per_month = FALSE
)
}
\arguments{
\item{year}{the year to produce plot for}

\item{df}{optinal; dataframe from the `sebms_sunhour_diff()`}

\item{month}{numeric value of the months to summarise sun ours over (default to 4:9)}

\item{per_month}{logical; summarise per month instead of per year}
}
\value{
a figure that shows diffeence in sunhours
}
\description{
Produce a plot that shows differences in sun hours between a given year and mean.
}
