% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sebms_sundiff_plot}
\alias{sebms_sundiff_plot}
\title{Create a Figure that Shows the Difference in Sunhours}
\usage{
sebms_sundiff_plot(year = year(today()) - 1, df, sunvar = diffsun, month = 4:9)
}
\arguments{
\item{year}{the year to compare with}

\item{df}{optinal; dataframe from the `sebms_sunhour_diff()`}

\item{sunvar}{the variable to calculate colours on, `diffsun` for now}

\item{month}{}
}
\value{
a figure that shows diffeence in sunhours
}
\description{
Produce a plot that shows differences in sun hours between a given year and mean.
}
