% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soltimmar.R
\name{sebms_sunhours_data}
\alias{sebms_sunhours_data}
\title{Create a total Sunhours Dataframe from SMHI Sunhour data}
\usage{
sebms_sunhours_data(
  year = year(today()) - 1,
  months = 4:9,
  per_month = FALSE,
  per_day = FALSE,
  to_env = FALSE
)
}
\arguments{
\item{year}{the year to produce plot for}

\item{months}{numeric value of the months to summarise sun ours over (default to 4:9)}

\item{per_month}{logical; summarise per month instead of per year}

\item{per_day}{logical; if data should be downloaded per day}

\item{to_env}{logical; also send the result to the global environment as an object called 'spatsunist_{year}. If the function is used within a plot function this is TRUE}

\item{day}{the day of interest}

\item{hour}{the hour of interest}
}
\value{
a sf spatial point object with the WGS84 coordinate system
}
\description{
Produce a data frame of the total sunhours in Sweden for the given month.
}
