% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLqueries.R
\name{sebms_trimobs}
\alias{sebms_trimobs}
\title{Get Species Observations for Trim Indices}
\usage{
sebms_trimobs(
  year = 2010:lubridate::year(lubridate::today()),
  Art = 1:200,
  Län = ".",
  Region = ".",
  Landskap = ".",
  Kommun = ".",
  filterPattern = NULL,
  minmax = 22:32,
  source = c(54, 55, 56, 63, 64, 66, 67, 84)
)
}
\arguments{
\item{year}{year or years to use for plot}

\item{Art}{integer; the species id of interest}

\item{filterPattern}{a regex pattern to filter SQL query}

\item{minmax}{the first and last week of interest}

\item{source}{the database sources as id numbers,
defaults to `54,55,56,63,64,66,67,84`}
}
\value{
a tibble with number of observed individuals per year and site.
}
\description{
Retrieve data with observations of individuals per year.
}
